Rename providers/fake-provider.ts to providers/fake-ai.ts
